SQUASH: Address @nbokulich's review
