--- q2_sample_classifier/tests/data/vaw.map.txt
@@ -1,12 +1,0 @@
-#SampleID	Column
-GG_OTU_1	a
-GG_OTU_2	a
-GG_OTU_3	a
-GG_OTU_4	a
-GG_OTU_5	a
-Sample1	b
-Sample2	b
-Sample3	b
-Sample4	b
-Sample5	b
-Sample6	b
