+++ q2_sample_classifier/tests/data/vaw_importance.tsv
@@ -3,9 +3,3 @@ GG_OTU_2	0.07760118417569697
 GG_OTU_3	0.06570251750505914
 GG_OTU_4	0.061718558716901406
 GG_OTU_5	0.028086160290024458
-Sample1	0.02610253907515516
-Sample2	0.025273854068462408
-Sample3	0.02029344879308132
-Sample4	0.017580266040786597
-Sample5	0.017034058634807364
-Sample6	0.016776104364241963
